* NEWS configure.in docs//* include/libvirt/libvirt.h libvirt.specx.*:   preparing release 0.1.5 * proxy/libvirt_proxy.c src/xen_internal.c: tiny fixes Daniel

+++ proxy/libvirt_proxy.c
@@ -576,7 +576,7 @@ retry2:
                 req->len = sizeof(virProxyPacket);
 	    } else {
                 int xmllen = strlen(xml);
-                if (xmllen > sizeof(request.extra.str)) {
+                if (xmllen > (int) sizeof(request.extra.str)) {
                     req->data.arg = -2;
                     req->len = sizeof(virProxyPacket);
                 } else {
